requirements.txt
pinecone-client==2.2.1
Flask==2.2.2
openai==0.27.6
Flask-Cors==3.0.10
